+++ docs/changelog.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## Unreleased
+## v0.11.0
 - add `modlog status unlogged` to view unlogged modlog events
 - add various options to control the output of the `.roles` command
 - add support for assigning multiple roles at once in `.assign` command
 - add the `.role mute` command to configure the role to be applied on member mute
+- add the `.mute`, `.tempmute` and `.unmute` commands
 
 ## v0.10.2
 - prevent duplicate `add role` join actions on a per-role basis